backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["AsgrimS <jan.semik@gmail.com>"]
readme = "README.md"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.0.0"
requests = "^2.31.0"
pyjwt = "^2.7.0"
cryptography = "^41.0.1"
markdown = "^3.4.3"
authlib = "^1.2.0"

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
  "E", # PEP8
  "W", # PEP8
  "N", # PEP8-Naming
  "F", # Pyflakes
  "I", # Isort
]
unfixable = ["F401"]


[tool.black]
line-length = 100
target-version = ["py311"]


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.270"
black = "^23.3.0"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
